requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
streamlit
python-dotenv
ipykernel
pyarrow
fastparquet